+++ backend/Backend.Domain/ValueObjects/Price.cs
@@ -2,6 +2,6 @@
 
 public record Price
 {
-    public required double Initial { get; init; }
-    public required double Discounted { get; init; }
+    public required decimal Initial { get; init; }
+    public required decimal Discounted { get; init; }
 }
